플레이덱매니저와 플레이덱매니저모델 기능분리
