+++ CardGameApp/CardGameApp/views/PlayDeckViewManager.swift
@@ -10,21 +10,8 @@ import UIKit
 
 /// 플레이덱뷰 를 가지는 객체
 class PlayDeckViewManager : UIView{
-    /// 각 라인덱을 배열로 같는다
-    var deckList : [PlayDeckView] = []
-    
-    /// 카드사이즈와 좌표를 받아서 스택뷰 구성
-    private func makeDeckViews(cardSize: CardSize, xPositions: [CGFloat], yPositions: [CGFloat]){
-        for count in 0..<cardSize.maxCardCount {
-            // 스택뷰 구성
-            let stackView = PlayDeckView()
-            stackView.setting(cardSize: cardSize, x: xPositions[count], y: yPositions[0])
-            
-            // 일반추가함
-            self.addSubview(stackView)
-            deckList.append(stackView)
-        }
-    }
+    /// 매니저 선언
+    var playDeckViewManagerModel = PlayDeckViewManagerModel()
     
     /// 플레이덱매니저 초기세팅함수
     func setting(cardSize: CardSize, xPositions: [CGFloat], yPositions: [CGFloat]){
@@ -34,20 +21,25 @@ class PlayDeckViewManager : UIView{
         self.frame.size.width = cardSize.screenWidth
         self.frame.size.height = cardSize.screenHeight - yPositions[1]
         // 내부 덱뷰를 생성한다
-        makeDeckViews(cardSize: cardSize, xPositions: xPositions, yPositions: yPositions)
+        let deckList = playDeckViewManagerModel.setting(cardSize: cardSize, xPositions: xPositions, yPositions: yPositions)
+        // 생성된 덱뷰를 서브뷰로 추가한다
+        addPlayDeckViews(playDeckViews: deckList)
+    }
+    
+    /// 플레이덱뷰 배열을 받아서 서브뷰로 추가
+    private func addPlayDeckViews(playDeckViews: [PlayDeckView]){
+        for playDeckView in playDeckViews {
+            self.addSubview(playDeckView)
+        }
     }
     
     /// 덱라인, 뷰를 받아서 추가
     func addView(view: UIView, deckLine: Int){
-        // 덱 라인에 맞는 스택뷰 추출
-        let stackView = self.deckList[deckLine]
-        // 맞는 스택뷰에 추가
-        stackView.addPlayCardview(view)
+        self.playDeckViewManagerModel.addView(view: view, deckLine: deckLine)
     }
     
     /// 과거카드데이터를 받아서 해당 뷰 리턴
     func getView(pastCardData: PastCardData) -> UIView? {
-        return self.deckList[pastCardData.deckLine].lastView()
-        
+        return self.playDeckViewManagerModel.getView(pastCardData: pastCardData)
     }
 }
